code/requirements.txt
openai>=1.0
anthropic>=0.20
datasets
tiktoken
nltk
tqdm
pandas
python-dotenv
